updates and conveniences
